requirements.txt
azure-storage-blob
azure-ai-documentintelligence
azure-search-documents
openai
python-dotenv
pandas
streamlit
